Memory corruption when server unreachable, and multiple get segfaults
`mmc_pool_select` change:
The problem was that memcache.so will execute “get()s” when “set()s” are
requested and vice versa. This would happen because in certain cases of
Memcached failures, instead of retrying failed requests memcache.so will
just forget about them and move on to the next request, while leaving
the state of the old request uncleaned. Then when a new request comes,
memcache.so will retry the previously failed request which already has
many of it’s pointers free()d and thus the segfaults.

`mmc_pool_run` changes:
Executing memcache_get()s with multiple keys were not retried properly
and left for the next request, where they segfaulted.